Bug 1121956 - Update response handling to cope with ElasticSearch v1.x
ElasticSearch v1.x no longer returns 'ok', due to:
https://github.com/elasticsearch/elasticsearch/commit/d23f640cd1488727c8337a888527d662c6135fac

Ideally we should be checking the HTTP response code, but bzcache uses
mozautoeslib, which in turn uses an ancient version of pyes - and it's not worth
delving into that rabbit hole. This solution isn't elegant, but should work for
now until OrangeFactor is end of lifed.