Bug 1286935 - Stop dropping the bzcache index on each import
Since it increases the problems if later imports fail, plus gives a
window when bzcache is empty, causing OF requests to fallback to
Bugzilla in a big way.

Stale entries in the bzcache index shouldn't be a problem - very few
intermittent bugs ever have their keyword removed, and for those that
do, they'll just be ignored.